python-face-service/requirements.txt
fastapi
uvicorn[standard]
face_recognition
numpy
pillow
